:construction: Remove references to Redux
