--- packages/connect-wallet/src/hooks/useAppState.ts
@@ -1,6 +1,0 @@
-import {TypedUseSelectorHook, useDispatch, useSelector} from 'react-redux';
-
-import type {RootState, AppDispatch} from '~/store/configureStore';
-
-export const useAppDispatch = () => useDispatch<AppDispatch>();
-export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
