--- packages/connect-wallet/src/slices/modalSlice/index.ts
@@ -1,1 +1,0 @@
-export * from './modalSlice';
