--- packages/connect-wallet/src/slices/walletSlice/gateContextClient.ts
@@ -1,12 +1,0 @@
-import {
-  getGateContextClient,
-  undefinedGateContextGenerator,
-} from '@shopify/gate-context-client';
-
-const defaultClientBuilder = () =>
-  getGateContextClient({
-    backingStore: 'ajaxApi',
-    shopifyGateContextGenerator: undefinedGateContextGenerator,
-  });
-
-export const gateContextClient = defaultClientBuilder();
