--- packages/connect-wallet/src/slices/walletSlice/index.ts
@@ -1,4 +1,0 @@
-export * from './walletSlice';
-export {attributeOrder} from './attributeOrder';
-export {fetchEns} from './fetchEns';
-export {fetchDelegations} from './delegateCash';
